+++ src/routes/auth.py
@@ -13,7 +13,7 @@ from src.models import (
     UserDashboardSettings,
 )
 from src.utils import render_template_from_file, ROOT_DIR
-from src.routes.helper import get_email_addresses
+from src.routes.helper.common_helper import get_email_addresses
 from src.config import get_app_info
 
 auth_bp = blueprints.Blueprint("auth", __name__)
