+++ src/routes/graphs.py
@@ -1,11 +1,11 @@
 from flask import render_template, blueprints
 from flask_login import login_required
 from src.config import app
+from src.routes.helper.common_helper import admin_required
 
 graphs_bp = blueprints.Blueprint("graphs", __name__)
 
-
 @app.route('/graphs')
-@login_required
+@admin_required
 def graphs():
     return render_template('graphs/graphs.html')
