+++ src/routes/ping.py
@@ -1,10 +1,8 @@
 import datetime
-from threading import Timer
 from flask import render_template, request, redirect, url_for, blueprints, flash
 from src.config import app, db
 from src.models.monitored_website import MonitoredWebsite
-from flask_login import login_required
-import requests
+from src.routes.helper.common_helper import admin_required
 
 ping_bp = blueprints.Blueprint('ping', __name__)
 
@@ -38,7 +36,7 @@ def time_ago(ping_time, current_time):
 
 # Route to view and add websites
 @app.route('/monitor_websites')
-@login_required
+@admin_required
 def monitor_websites():
     websites = MonitoredWebsite.query.all()
     current_time = datetime.datetime.now()
@@ -48,7 +46,7 @@ def monitor_websites():
 
 # Route to add a website
 @app.route('/add_monitored_website', methods=['POST'])
-@login_required
+@admin_required
 def add_website():
     name = request.form['name']
     ping_interval = int(request.form['ping_interval'])
@@ -63,7 +61,7 @@ def add_website():
     return redirect(url_for('monitor_websites'))
 
 @app.route('/delete_monitored_website/<int:website_id>', methods=['POST'])
-@login_required
+@admin_required
 def remove_website(website_id):
     website = MonitoredWebsite.query.get_or_404(website_id)
     db.session.delete(website)
@@ -72,6 +70,7 @@ def remove_website(website_id):
     return redirect(url_for('monitor_websites'))
 
 @app.route('/edit_monitored_website/<int:website_id>', methods=['GET', 'POST'])
+@admin_required
 def edit_website(website_id):
     website = MonitoredWebsite.query.get_or_404(website_id)
     if request.method == 'POST':
@@ -86,7 +85,7 @@ def edit_website(website_id):
 
 # Route to toggle pinging status
 @app.route('/toggle_ping_status/<int:website_id>')
-@login_required
+@admin_required
 def toggle_ping(website_id):
     website = MonitoredWebsite.query.get_or_404(website_id)
     website.is_ping_active = not website.is_ping_active
@@ -94,7 +93,7 @@ def toggle_ping(website_id):
     return redirect(url_for('monitor_websites'))
 
 @app.route('/toggle_email_alerts/<int:website_id>')
-@login_required
+@admin_required
 def toggle_email_alerts(website_id):
     website = MonitoredWebsite.query.get_or_404(website_id)
     website.email_alerts_enabled = not website.email_alerts_enabled
