+++ src/routes/smtp_email_config.py
@@ -3,12 +3,13 @@ from flask import render_template, request, flash, redirect, url_for, blueprints
 
 from src.config import app, db
 from src.models import SMTPSettings
+from src.routes.helper.common_helper import admin_required
 
 smtp_email_config_bp = blueprints.Blueprint('smtp_email_config', __name__)
 
 
 @app.route("/update-email-password", methods=["GET", "POST"])
-@login_required
+@admin_required
 def smtp_config():
     smtp_config = SMTPSettings.query.first()
