+++ src/static/css/control_tab.css
@@ -4,6 +4,7 @@
     padding: 30px;
     border-radius: 15px;
     box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
+    margin-top: 50px;
 }
 
 .settings-title {
