+++ src/templates/ext/navbar.html
@@ -74,14 +74,12 @@
                 </a>
             </li>
             {% endif %}
-            {% if current_user.user_level == 'admin' %}
             <li class="nav-item">
                 <a class="nav-link {% if request.endpoint == 'settings' %}active{% endif %}"
                     href="{{ url_for('settings') }}">
                     <i class="fas fa-cog mr-1"></i>Settings
                 </a>
             </li>
-            {% endif %}
         </ul>
         <!-- Authentication Links on the Right Side -->
         <ul class="navbar-nav ml-auto">
